move nin types to it's own file
